Enable hour values > 23 for timers
Add timertime(), similar to localtime(3), which takes a time_t in
and returns a struct tm filled only with secs, mins and unlimited hours

add timerformat(), a wrapper for strftime which checks if the specified
format starts with '%H' and simply writes hours with sprintf() then to
avoid limiting them to 23. Afterwards and if the format doesn't start
with '%H' strftime() is called with the remaining output and format.