Add option to daemonize timer
-d will make it so the process daemonizes, prints its PID and timer
length and takes the steps to notify the user of its end after the
required time has passed. Because there is no need to animate a clock
here, we can elegantly nanosleep the distance without any
calculations, respect for pausing, etc.

Also couple SDL init and playing steps (which were separate in the
option handling and playing phases) into a function which can be used
when necessary. (This was necessary because initializing and executing
SDL in different processes obviously doesn't work. But there are also
no huge drawbacks to initializing SDL when trying to play. Except that
the function can't be used twice, I just realized. We don't do this, but
I will add a fixme nonetheless.)